+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Dish classifier service ([#31](https://github.com/kooznitsa/sampo/pull/31)).
+- Dish tags generation on save ([#32](https://github.com/kooznitsa/sampo/pull/32)).
 
 ### Fixed