+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dish tags generation commands ([#33](https://github.com/kooznitsa/sampo/pull/33)).
 - Admin methods to create restaurant nearest stations ([#34](https://github.com/kooznitsa/sampo/pull/34)).
 - Filter list of dishes by tags ([#36](https://github.com/kooznitsa/sampo/pull/36)).
+- Production settings ([#38](https://github.com/kooznitsa/sampo/pull/38)).
 
 ### Fixed