+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Permission limitations for GET requests ([#40](https://github.com/kooznitsa/sampo/pull/40)).
+
 ## [1.2.1] - 2025-11-18
 
 ### Added